Angles

Type for easily working with angles in DMS format and converting to/from decimal.

CelestialCalc.Angle — Type

Angle

Represents an angle with degrees, minutes and seconds. The isnegative flag is used instead of allowing degrees to be negative for better dealing with negative angles smaller than 1 (example: -0°10'10''). For convenience, the constructors accept negative degrees, but the final object always has positive degrees and sets the isnegative as needed.

Examples

julia> Angle(10,5,30)
10°05'30.00''

julia> Angle(45,15,2,true)
-45°15'02.00''

julia> Angle(-45,15,2)
-45°15'02.00''

julia> Angle(270,55,2.1)
270°55'02.10''

CelestialCalc.Angle — Type

Angle(d, isneg::Bool=false) -> Angle

Construct an angle with whole degrees.

Examples

julia> Angle(10)
10°00'00.00''

julia> Angle(45,true)
-45°00'00.00''

CelestialCalc.Angle — Type

Angle(d, m, isneg::Bool=false) -> Angle

Construct an angle with whole degrees and minutes.

Examples

julia> Angle(10,5)
10°05'00.00''

julia> Angle(45,15,true)
-45°15'00.00''

CelestialCalc.Angle — Method

Angle(d, m, s::Real) -> Angle

Shorthand for the construction of positive angles.

Examples

julia> Angle(10,5,30)
10°05'30.00''

julia> Angle(45,15,2.1)
45°15'02.10''

CelestialCalc.angle_to_decimal — Method

angle_to_decimal(angle::Angle)

Convert Angle in DMS format to a decimal.

CelestialCalc.decimal_to_angle — Method

decimal_to_angle(decimal::Real)

Convert a decimal value to an Angle in DMS format.