Coordinate Systems

Types for the positional representation of observers and sky objects, and functions for converting between them.

CelestialCalc.LatLng — Type

LatLng

Coordinate system for locating a point in Earth's surface.

CelestialCalc.EquatorialCoordinates — Type

EquatorialCoordinates

Coordinate system for locating a point in the celestial sphere, given fixed points of reference (the celestial equator and the First Point of Aries). Here, α is right-ascention and δ is declination.

CelestialCalc.HorizonCoordinates — Type

HorizonCoordinates

Coordinate system for locating a point in an observer's local celestial sphere, which varies with time. Here, h is altitude and Az is azimuth.

CelestialCalc.equatorial_to_horizon — Function

equatorial_to_horizon(δ::Float64, hour_angle::Time, latitude::Float64) -> HorizonCoordinates

Find the HorizonCoordinates for an object, given its declination (δ) and hour angle (H) represented as a Time object, as well as the observer's latitude.

equatorial_to_horizon(δ::Float64, hour_angle::Float64, latitude::Float64) -> HorizonCoordinates

Find the HorizonCoordinates for an object, given its declination (δ) and hour angle (H) represented in decimal format, as well as the observer's latitude.

equatorial_to_horizon(eqcoord::EquatorialCoordinates, local_civilian_date::ZonedDateTime, latlong::LatLng) -> HorizonCoordinates

Find the HorizonCoordinates for an object, given its EquatorialCoordinates and the observer's local civilizan time and position.